+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 - Improve internal logging capabilities
 - Add link to GitHub issues project page to footer
 
+## Changed
+- Remember form values after bad signup (name already in use, bad email etc.)
+
 ## 2018-08-24
 ### Added
 - 'Like' button added for expression of joy and satisfaction over a post